+++ src/version2/appProperties.ts
@@ -147,11 +147,74 @@ export class AppProperties {
   }
 
   /**
+   * Returns all property keys for the Forge app.
+   *
+   * **[Permissions](https://developer.atlassian.com/cloud/jira/platform/rest/v2/intro/#permissions) required:** Only
+   * Forge apps can make this request. This API can only be accessed using
+   * **[asApp()](https://developer.atlassian.com/platform/forge/apis-reference/fetch-api-product.requestjira/#method-signature)**
+   * requests from Forge.
+   */
+  async getForgeAppPropertyKeys<T = Models.GetForgeAppPropertyKeys>(callback: Callback<T>): Promise<void>;
+  /**
+   * Returns all property keys for the Forge app.
+   *
+   * **[Permissions](https://developer.atlassian.com/cloud/jira/platform/rest/v2/intro/#permissions) required:** Only
+   * Forge apps can make this request. This API can only be accessed using
+   * **[asApp()](https://developer.atlassian.com/platform/forge/apis-reference/fetch-api-product.requestjira/#method-signature)**
+   * requests from Forge.
+   */
+  async getForgeAppPropertyKeys<T = Models.GetForgeAppPropertyKeys>(callback?: never): Promise<T>;
+  async getForgeAppPropertyKeys<T = Models.GetForgeAppPropertyKeys>(callback?: Callback<T>): Promise<void | T> {
+    const config: RequestConfig = {
+      url: '/rest/forge/1/app/properties',
+      method: 'GET',
+    };
+
+    return this.client.sendRequest(config, callback);
+  }
+
+  /**
+   * Returns the value of a Forge app's property.
+   *
+   * **[Permissions](https://developer.atlassian.com/cloud/jira/platform/rest/v2/intro/#permissions) required:** Only
+   * Forge apps can make this request. This API can only be accessed using
+   * **[asApp()](https://developer.atlassian.com/platform/forge/apis-reference/fetch-api-product.requestjira/#method-signature)**
+   * requests from Forge.
+   */
+  async getForgeAppProperty<T = Models.GetForgeAppProperty>(
+    parameters: Parameters.GetForgeAppProperty,
+    callback: Callback<T>,
+  ): Promise<void>;
+  /**
+   * Returns the value of a Forge app's property.
+   *
+   * **[Permissions](https://developer.atlassian.com/cloud/jira/platform/rest/v2/intro/#permissions) required:** Only
+   * Forge apps can make this request. This API can only be accessed using
+   * **[asApp()](https://developer.atlassian.com/platform/forge/apis-reference/fetch-api-product.requestjira/#method-signature)**
+   * requests from Forge.
+   */
+  async getForgeAppProperty<T = Models.GetForgeAppProperty>(
+    parameters: Parameters.GetForgeAppProperty,
+    callback?: never,
+  ): Promise<T>;
+  async getForgeAppProperty<T = Models.GetForgeAppProperty>(
+    parameters: Parameters.GetForgeAppProperty,
+    callback?: Callback<T>,
+  ): Promise<void | T> {
+    const config: RequestConfig = {
+      url: `/rest/forge/1/app/properties/${parameters.propertyKey}`,
+      method: 'GET',
+    };
+
+    return this.client.sendRequest(config, callback);
+  }
+
+  /**
    * Sets the value of a Forge app's property. These values can be retrieved in [Jira
    * expressions](https://developer.atlassian.com/cloud/jira/platform/jira-expressions/) through the `app` [context
    * variable](https://developer.atlassian.com/cloud/jira/platform/jira-expressions/#context-variables). They are also
    * available in [entity property display
-   * conditions](/platform/forge/manifest-reference/display-conditions/entity-property-conditions/).
+   * conditions](https://developer.atlassian.com/platform/forge/manifest-reference/display-conditions/entity-property-conditions/).
    *
    * For other use cases, use the [Storage
    * API](https://developer.atlassian.com/platform/forge/runtime-reference/storage-api/).