+++ src/version2/models/bulkWorklogKeyRequest.ts
@@ -1,0 +1,6 @@
+import type { WorklogCompositeKey } from './worklogCompositeKey';
+
+export interface BulkWorklogKeyRequest {
+  /** A list of issue and worklog ID pairs. */
+  requests?: WorklogCompositeKey[];
+}