+++ src/version2/models/fieldAssociationSchemeLinks.ts
@@ -1,0 +1,4 @@
+export interface FieldAssociationSchemeLinks {
+  associations?: string;
+  projects?: string;
+}