+++ src/version2/models/fieldAssociationSchemeProjectSearchResult.ts
@@ -1,0 +1,5 @@
+/** Project search results for field association scheme. */
+export interface FieldAssociationSchemeProjectSearchResult {
+  id?: string;
+  name?: string;
+}