+++ src/version2/models/searchResultFieldParameters.ts
@@ -1,0 +1,4 @@
+export interface SearchResultFieldParameters {
+  description?: string;
+  isRequired?: boolean;
+}