+++ src/version2/models/workTypeParameters.ts
@@ -1,0 +1,5 @@
+export interface WorkTypeParameters {
+  description?: string;
+  isRequired: boolean;
+  workTypeId: number;
+}