+++ src/version2/models/worklogCompositeKey.ts
@@ -1,0 +1,6 @@
+export interface WorklogCompositeKey {
+  /** The issue ID. */
+  issueId?: number;
+  /** The worklog ID. */
+  worklogId?: number;
+}