+++ src/version2/parameters/parameterRemovalDetails.ts
@@ -1,0 +1,8 @@
+export interface ParameterRemovalDetails {
+  /** Set of parameter names to remove */
+  parameters?: string[];
+  /** ID of the field scheme */
+  schemeId?: number;
+  /** Set of work type (issue type) IDs */
+  workTypeIds?: number[];
+}