+++ src/version2/parameters/switchWorkflowSchemeForProject.ts
@@ -1,0 +1,3 @@
+import type { WorkflowSchemeProjectSwitch } from '../models';
+
+export type SwitchWorkflowSchemeForProject = WorkflowSchemeProjectSwitch;